更新

+++ src/main/java/com/github/julyss2019/mcsp/julyguild/log/guild/GuildBalanceChangedLog.java
@@ -4,19 +4,20 @@ import com.github.julyss2019.mcsp.julyguild.guild.GuildBank;
 import com.github.julyss2019.mcsp.julyguild.log.BaseGuildLog;
 import com.github.julyss2019.mcsp.julyguild.log.GuildLogType;
 
+import java.math.BigDecimal;
+
 public class GuildBalanceChangedLog extends BaseGuildLog {
     private GuildBank.BalanceType balanceType;
-    private Double oldBalance;
-    private Double newBalance;
+    private BigDecimal oldBalance;
+    private BigDecimal newBalance;
 
-    public GuildBalanceChangedLog(String guildUUID, GuildBank.BalanceType balanceType, Double oldBalance, Double newBalance) {
+    public GuildBalanceChangedLog(String guildUUID, GuildBank.BalanceType balanceType, BigDecimal oldBalance, BigDecimal newBalance) {
         super(GuildLogType.BALANCE_CHANGED, guildUUID);
 
         this.balanceType = balanceType;
         this.oldBalance = oldBalance;
         this.newBalance = newBalance;
     }
-
 
     public GuildBank.BalanceType getBalanceType() {
         return balanceType;
@@ -26,19 +27,19 @@ public class GuildBalanceChangedLog extends BaseGuildLog {
         this.balanceType = balanceType;
     }
 
-    public Double getOldBalance() {
+    public BigDecimal getOldBalance() {
         return oldBalance;
     }
 
-    public void setOldBalance(Double oldBalance) {
+    public void setOldBalance(BigDecimal oldBalance) {
         this.oldBalance = oldBalance;
     }
 
-    public Double getNewBalance() {
+    public BigDecimal getNewBalance() {
         return newBalance;
     }
 
-    public void setNewBalance(Double newBalance) {
+    public void setNewBalance(BigDecimal newBalance) {
         this.newBalance = newBalance;
     }
 }
